// frontend/src/app/core/models/ai-results-response.model.ts
export class AiResultsResponse {
  id?: any;
  studentName?: string;
  studentId?: number;
  studentRollNumber?: any;
  studentEmail?: string;
  className?: any;
  assessmentName?: any;
  grade?: number;
  score?: number;
  confidenceLevel?: string;
  creationDate?: string;
  assignmentId?: number;
  assignmentTitle?: string;
  resultStatus?: string;
  status?: string;
  emailSent?: boolean;
  fileUrl?: string;
  isEditingEmail: any;
  valuesLoader?: boolean;
}